DataLoopNode: change constexpr to enum 4

+++ src/EnergyPlus/AirflowNetworkBalanceManager.cc
@@ -9658,7 +9658,7 @@ namespace AirflowNetworkBalanceManager {
             } else {
                 //   skip nodes for air cooled condensers
                 for (j = 1; j <= isize(NodeConnectionType); ++j) {
-                    if (NodeConnectionType(j) == DataLoopNode::NodeConnectionType::OutsideAirReference) {
+                    if (NodeConnectionType(j) == static_cast<int>(DataLoopNode::NodeConnectionType::OutsideAirReference)) {
                         NodeFound(i) = true;
                     }
                 }
